Hardcore changes, emi whitelist

+++ src/main/java/com/livajq/arcanetweaks/mixin/vanilla/LivingEntityMixin.java
@@ -36,8 +36,8 @@ public class LivingEntityMixin implements TickSpeedAccessor {
     private float redirectResCalc(float a, float b, DamageSource src, float pDamageAmount) {
         LivingEntity self = (LivingEntity)(Object)this;
         
-        final float resPerLvl = Config.resistanceAmount; //% reduction per resistance level
-        pDamageAmount -= pDamageAmount * (self.getEffect(MobEffects.DAMAGE_RESISTANCE).getAmplifier() + 1) * resPerLvl;
+        final double resPerLvl = Config.resistanceAmount; //% reduction per resistance level
+        pDamageAmount -= (float) (pDamageAmount * (self.getEffect(MobEffects.DAMAGE_RESISTANCE).getAmplifier() + 1) * resPerLvl);
         return Math.max(pDamageAmount, 0.0F);
     }
 }
